use pipeline incrementally to prevent context mismatch on async activities
