+++ posthog/warehouse/data_load/pipeline.py
@@ -1,28 +1,32 @@
 from dataclasses import dataclass
-from typing import Dict, List
+from typing import Dict
 
 import dlt
 import s3fs
-from asgiref.sync import sync_to_async
 from django.conf import settings
 from dlt.pipeline.exceptions import PipelineStepFailed
 
 from posthog.warehouse.models import ExternalDataSource
 
-from .stripe import ENDPOINTS, stripe_source
+from .stripe import ENDPOINTS, stripe_pagination
+
 import os
+import stripe
+
+stripe.api_version = "2022-11-15"
 
 
 @dataclass
 class PipelineInputs:
+    source_id: str
     dataset_name: str
     job_type: str
     team_id: int
 
 
 def create_pipeline(inputs: PipelineInputs):
-    pipeline_name = f"{inputs.job_type}_pipeline_{inputs.team_id}"
-    pipelines_dir = f"{os.getcwd()}/.dlt/{inputs.team_id}/{inputs.job_type}"
+    pipeline_name = f"{inputs.job_type}_pipeline_{inputs.team_id}_source_{inputs.source_id}"
+    pipelines_dir = f"{os.getcwd()}/.dlt/{inputs.team_id}/{inputs.source_id}/{inputs.job_type}"
     return dlt.pipeline(
         pipeline_name=pipeline_name,
         pipelines_dir=pipelines_dir,  # workers can be created and destroyed so it doesn't matter where the metadata gets put temporarily
@@ -58,50 +62,16 @@ class StripeJobInputs(PipelineInputs):
 PIPELINE_TYPE_SCHEMA_DEFAULT_MAPPING = {ExternalDataSource.Type.STRIPE: ENDPOINTS}
 
 
-@sync_to_async
-def run_stripe_pipeline(inputs: StripeJobInputs) -> List[SourceSchema]:
-    pipeline = create_pipeline(inputs)
-
-    # TODO: decouple API calls so they can be incrementally read and sync_rows updated
-    source = stripe_source(
-        stripe_secret_key=inputs.stripe_secret_key,
-        endpoints=PIPELINE_TYPE_SCHEMA_DEFAULT_MAPPING[ExternalDataSource.Type.STRIPE],
-    )
-    try:
-        pipeline.run(source, loader_file_format="parquet")
-    except PipelineStepFailed:
-        # TODO: log
-        raise
-
-    return get_schema(pipeline)
-
-
-def get_schema(pipeline: dlt.pipeline) -> List[SourceSchema]:
-    schema = pipeline.default_schema
-    data_tables = schema.data_tables()
-    schemas = []
-
-    for resource in data_tables:
-        columns = {}
-        try:
-            for column_name, column_details in resource["columns"].items():
-                columns[column_name] = SourceColumnType(
-                    name=column_details["name"],
-                    data_type=column_details["data_type"],
-                    nullable=column_details["nullable"],
-                )
-
-            resource_schema = SourceSchema(
-                resource=resource["resource"],
-                name=resource["name"],
-                columns=columns,
-                write_disposition=resource["write_disposition"],
-            )
-            schemas.append(resource_schema)
-        except:
-            pass
-
-    return schemas
+async def run_stripe_pipeline(inputs: StripeJobInputs) -> None:
+    for endpoint in ENDPOINTS:
+        async for item in stripe_pagination(inputs.stripe_secret_key, endpoint):
+            try:
+                pipeline = create_pipeline(inputs)
+                pipeline.run(item, table_name=endpoint.lower(), loader_file_format="parquet")
+                pipeline.deactivate()
+            except PipelineStepFailed:
+                # TODO: log
+                raise
 
 
 PIPELINE_TYPE_INPUTS_MAPPING = {ExternalDataSource.Type.STRIPE: StripeJobInputs}
