+++ posthog/warehouse/data_load/stripe/__init__.py
@@ -1,52 +1,73 @@
-""" This source uses Stripe API and dlt to load data such as Customer, Subscription, Event etc. to the database and to calculate the MRR and churn rate. """
+"""Stripe analytics source helpers"""
 
-from typing import Any, Dict, Generator, Iterable, Optional, Tuple
+from typing import Any, Dict, Iterable, Optional, Union
 
-import dlt
 import stripe
-from dlt.extract.source import DltResource
+from dlt.common import pendulum
+from dlt.common.typing import TDataItem
 from pendulum import DateTime
-
-from .helper import pagination
+from asgiref.sync import sync_to_async
 from .settings import ENDPOINTS
 
 
-@dlt.source
-def stripe_source(
-    endpoints: Tuple[str, ...] = ENDPOINTS,
-    stripe_secret_key: str = dlt.secrets.value,
-    start_date: Optional[DateTime] = None,
-    end_date: Optional[DateTime] = None,
-) -> Iterable[DltResource]:
+async def stripe_pagination(
+    api_key: str, endpoint: str, start_date: Optional[Any] = None, end_date: Optional[Any] = None
+) -> Iterable[TDataItem]:
     """
-    Retrieves data from the Stripe API for the specified endpoints.
-
-    For all endpoints, Stripe API responses do not provide the key "updated",
-    so in most cases, we are forced to load the data in 'replace' mode.
-    This source is suitable for all types of endpoints, including 'Events', 'Invoice', etc.
-    but these endpoints can also be loaded in incremental mode (see source incremental_stripe_source).
+    Retrieves data from an endpoint with pagination.
 
     Args:
-        endpoints (Tuple[str, ...]): A tuple of endpoint names to retrieve data from. Defaults to most popular Stripe API endpoints.
-        stripe_secret_key (str): The API access token for authentication. Defaults to the value in the `dlt.secrets` object.
-        start_date (Optional[DateTime]): An optional start date to limit the data retrieved. Format: datetime(YYYY, MM, DD). Defaults to None.
-        end_date (Optional[DateTime]): An optional end date to limit the data retrieved. Format: datetime(YYYY, MM, DD). Defaults to None.
+        endpoint (str): The endpoint to retrieve data from.
+        start_date (Optional[Any]): An optional start date to limit the data retrieved. Defaults to None.
+        end_date (Optional[Any]): An optional end date to limit the data retrieved. Defaults to None.
 
     Returns:
-        Iterable[DltResource]: Resources with data that was created during the period greater than or equal to 'start_date' and less than 'end_date'.
+        Iterable[TDataItem]: Data items retrieved from the endpoint.
     """
-    stripe.api_key = stripe_secret_key
-    stripe.api_version = "2022-11-15"
+    starting_after = None
+    while True:
+        response = await stripe_get_data(
+            api_key,
+            endpoint,
+            start_date=start_date,
+            end_date=end_date,
+            starting_after=starting_after,
+        )
 
-    def stripe_resource(
-        endpoint: str,
-    ) -> Generator[Dict[Any, Any], Any, None]:
-        for item in pagination(endpoint, start_date, end_date):
-            yield item
+        if len(response["data"]) > 0:
+            starting_after = response["data"][-1]["id"]
+        yield response["data"]
 
-    for endpoint in endpoints:
-        yield dlt.resource(
-            stripe_resource,
-            name=endpoint,
-            write_disposition="replace",
-        )(endpoint)
+        if not response["has_more"]:
+            break
+
+
+def transform_date(date: Union[str, DateTime, int]) -> int:
+    if isinstance(date, str):
+        date = pendulum.from_format(date, "%Y-%m-%dT%H:%M:%SZ")
+    if isinstance(date, DateTime):
+        # convert to unix timestamp
+        date = int(date.timestamp())
+    return date
+
+
+async def stripe_get_data(
+    api_key: str,
+    resource: str,
+    start_date: Optional[Any] = None,
+    end_date: Optional[Any] = None,
+    **kwargs: Any,
+) -> Dict[Any, Any]:
+    if start_date:
+        start_date = transform_date(start_date)
+    if end_date:
+        end_date = transform_date(end_date)
+
+    if resource == "Subscription":
+        kwargs.update({"status": "all"})
+
+    _resource = getattr(stripe, resource)
+    resource_dict = await sync_to_async(_resource.list)(
+        api_key=api_key, created={"gte": start_date, "lt": end_date}, limit=100, **kwargs
+    )
+    return dict(resource_dict)
